requirements.txt
blurhash==1.1.4
certifi==2019.9.11
chardet==3.0.4
decorator==4.4.0
idna==2.8
kafka-python==1.4.7
Mastodon.py==1.5.0
python-dateutil==2.8.0
python-magic==0.4.15
pytz==2019.3
requests==2.22.0
six==1.12.0
urllib3==1.25.6
